Separate out ModelVisualComponent
