+++ rootex/framework/components/transform_component.h
@@ -27,9 +27,7 @@ class TransformComponent : public Component
 	TransformComponent(const Vector3& position, const Vector4& rotation, const Vector3& scale);
 	TransformComponent(TransformComponent&) = delete;
 
-	friend class VisualComponent;
-	friend class DiffuseVisualComponent;
-	friend class VisualComponentAttributes;
+	friend class ModelVisualComponent;
 	friend class RenderSystem;
 	friend class EntityFactory;
